RESTful API 구현
